將 收件人資料的  元件加回去，調整資料互傳功能，再看哪裡需要再修正
